upgrade cfn response, add bill connect logic, add template output
