+++ src/Gemfile
@@ -1,2 +1,3 @@
 source 'https://rubygems.org'
-gem 'cfnresponse'
+gem 'cfn_response'
+gem 'faraday'
